reroutes: hardware -> dispatcher -> volume
