+++ shell/crates/hardware_buttons/src/handle.rs
@@ -19,11 +19,10 @@ pub fn build_message_for_event(event: KeyEvent) -> Option<Message> {
             println!("hardware_buttons: power button unknown event");
             message = Some(Message::ShowPowerOptions(true));
         }
-
         KeyEvent::Released(Key::Power) => {
             println!("hardware_buttons: power button released");
         }
-        
+
         KeyEvent::Pressed(Key::Home) => {
             println!("hardware_buttons: home button pressed");
         }
@@ -39,30 +38,61 @@ pub fn build_message_for_event(event: KeyEvent) -> Option<Message> {
 
         KeyEvent::Pressed(Key::VolumeUp) => {
             println!("hardware_buttons: volume up pressed");
+            message = Some(Message::VolumeUp);
         }
         KeyEvent::Pressing(Key::VolumeUp) => {
             println!("hardware_buttons: volume up pressing");
+            message = Some(Message::VolumeUp);
         }
         KeyEvent::Released(Key::VolumeUp) => {
             println!("hardware_buttons: volume up released");
         }
         KeyEvent::Unknown(Key::VolumeUp) => {
             println!("hardware_buttons: volume up unknown event");
+            message = Some(Message::VolumeUp);
         }
 
         KeyEvent::Pressed(Key::VolumeDown) => {
             println!("hardware_buttons: volume down pressed");
+            message = Some(Message::VolumeDown);
         }
         KeyEvent::Pressing(Key::VolumeDown) => {
             println!("hardware_buttons: volume down pressing");
+            message = Some(Message::VolumeDown);
         }
         KeyEvent::Released(Key::VolumeDown) => {
             println!("hardware_buttons: volume down released");
         }
         KeyEvent::Unknown(Key::VolumeDown) => {
             println!("hardware_buttons: volume down unknown event");
+            message = Some(Message::VolumeDown);
         }
-        _ => {}
+
+        KeyEvent::Pressed(Key::ExtensionDetection) => {
+            println!("hardware_buttons: extension detection pressed");
+        }
+        KeyEvent::Pressing(Key::ExtensionDetection) => {
+            println!("hardware_buttons: extension detection pressing");
+        }
+        KeyEvent::Released(Key::ExtensionDetection) => {
+            println!("hardware_buttons: extension detection released");
+        }
+        KeyEvent::Unknown(Key::ExtensionDetection) => {
+            println!("hardware_buttons: extension detection unknown event");
+        }
+
+        KeyEvent::Pressed(Key::Unknown) => {
+            println!("hardware_buttons: unknown key pressed");
+        }
+        KeyEvent::Pressing(Key::Unknown) => {
+            println!("hardware_buttons: unknown key pressing");
+        }
+        KeyEvent::Released(Key::Unknown) => {
+            println!("hardware_buttons: unknown key released");
+        }
+        KeyEvent::Unknown(Key::Unknown) => {
+            println!("hardware_buttons: unknown key unknown event");
+        }
     }
 
     message
